+++ CHANGELOG.md
@@ -10,6 +10,9 @@ For the full commit log, [see here](https://github.com/influxdata/influxdb-rails
   `ignored_environments` do disable instrumentation
 - Simplified spec with a PORO test client
 - Implement `instantiation.active_record` subscriber (https://guides.rubyonrails.org/active_support_instrumentation.html#instantiation-active-record)
+- Implement `enqueue.active_job` subscriber (https://guides.rubyonrails.org/active_support_instrumentation.html#enqueue-active-job)
+- Implement `perform_start.active_job` subscriber (https://guides.rubyonrails.org/active_support_instrumentation.html#perform-start-active-job)
+- Implement `perform.active_job` subscriber (https://guides.rubyonrails.org/active_support_instrumentation.html#perform-active-job)
 
 ## v1.0.0, released 2019-10-23
 The Final release, no code changes.